+++ CHANGELOG.md
@@ -15,12 +15,36 @@ Each version section may have have subsections for: _Added_, _Changed_, _Removed
 
 - Basic support for target factors in training.
 
+## [2.2.6]
+
+### Fixed
+- Avoid a crash due to an edge case when no model improvement has been observed by the time the learning rate gets reduced for the first time.
+
+## [2.2.5]
+
+### Fixed
+- Enforce sentence batching for sockeye score tool, set default batch size to 56
+
+## [2.2.4]
+
+### Changed
+- Use softmax with length in DotAttentionCell.
+- Use `contrib.arange_like` in AutoRegressiveBias block to reduce number of ops.
+
+## [2.2.3]
+
+### Added
+
+- Log the absolute number of `<unk>` tokens in source and target data
+
 ## [2.2.2]
 
 ### Fixed
- - Fix: Guard against null division for small batch sizes.
+
+- Fix: Guard against null division for small batch sizes.
 
 ## [2.2.1]
+
 ## Fixed
 
 - Fixes a corner case bug by which the beam decoder can wrongly return a best hypothesis with -infinite score.